Do not expose ports that are not used by default
Historically the image ran as root by default, so it could use root-only ports 80 and 443, but since the image runs as non-root by default now, we cannot run it easily on those ports, so they should not be exposed. Exposing ports that are not used can trick services that depend on them, like OpenShift, that uses the export ports to automatic route creation.