[obsolete code cleanup] Remove 'rpm -V' calls
The result of a discussion with the sclorg containers manitainers is that the 'rpm -V' is a remnant from the ages of 'yum', the 'dnf' predecessor
When 'yum' had to install a list of packages and some were not found, it happily installed only those found, not raising an error.
The 'rpm -V' here was a protection against that - a simple verification that all requested packages were actually found and installed.